+++ scripts/generate_cl.js
@@ -67,23 +67,6 @@ function rx()
 }
 
 
-function astrobwt()
-{
-    const astrobwt = opencl_minify(addIncludes('astrobwt.cl', [ 'BWT.cl', 'salsa20.cl', 'sha3.cl' ]));
-
-    // fs.writeFileSync('astrobwt_gen.cl', astrobwt);
-    fs.writeFileSync('astrobwt_cl.h', text2h(astrobwt, 'xmrig', 'astrobwt_cl'));
-}
-
-
-function astrobwt_v2()
-{
-    const astrobwt = opencl_minify(addIncludes('astrobwt_v2.cl', [ 'BWT.cl', 'salsa20.cl', 'sha3.cl' ]));
-
-    fs.writeFileSync('astrobwt_v2_cl.h', text2h(astrobwt, 'xmrig', 'astrobwt_v2_cl'));
-}
-
-
 function kawpow()
 {
     const kawpow = opencl_minify(addIncludes('kawpow.cl', [ 'defs.h' ]));
@@ -114,16 +97,6 @@ process.chdir(cwd);
 process.chdir(path.resolve('src/backend/opencl/cl/rx'));
 
 rx();
-
-process.chdir(cwd);
-process.chdir(path.resolve('src/backend/opencl/cl/astrobwt'));
-
-astrobwt();
-
-process.chdir(cwd);
-process.chdir(path.resolve('src/backend/opencl/cl/astrobwt_v2'));
-
-astrobwt_v2();
 
 process.chdir(cwd);
 process.chdir(path.resolve('src/backend/opencl/cl/kawpow'));