+++ src/backend/cpu/platform/BasicCpuInfo.cpp
@@ -379,16 +379,6 @@ xmrig::CpuThreads xmrig::BasicCpuInfo::threads(const Algorithm &algorithm, uint3
     }
 #   endif
 
-#   ifdef XMRIG_ALGO_ASTROBWT
-    if (f == Algorithm::ASTROBWT) {
-        CpuThreads threads;
-        for (size_t i = 0; i < count_limit; ++i) {
-            threads.add(i, 0);
-        }
-        return threads;
-    }
-#   endif
-
 #   ifdef XMRIG_ALGO_GHOSTRIDER
     if (f == Algorithm::GHOSTRIDER) {
         return CpuThreads(std::max<size_t>(count / 2, 1), 8);