+++ src/backend/cuda/CudaConfig_gen.h
@@ -138,22 +138,6 @@ size_t inline generate<Algorithm::RANDOM_X>(Threads<CudaThreads> &threads, const
 #endif
 
 
-#ifdef XMRIG_ALGO_ASTROBWT
-template<>
-size_t inline generate<Algorithm::ASTROBWT>(Threads<CudaThreads> &threads, const std::vector<CudaDevice> &devices)
-{
-    size_t count = 0;
-
-    if (!threads.isExist(Algorithm::ASTROBWT_DERO_2)) {
-        count += threads.move(Algorithm::kASTROBWT_DERO_2, CudaThreads(devices, Algorithm::ASTROBWT_DERO_2));
-    }
-
-    count += generate(Algorithm::kASTROBWT, threads, Algorithm::ASTROBWT_DERO, devices);
-    return count;
-}
-#endif
-
-
 #ifdef XMRIG_ALGO_KAWPOW
 template<>
 size_t inline generate<Algorithm::KAWPOW>(Threads<CudaThreads> &threads, const std::vector<CudaDevice> &devices)