+++ src/backend/opencl/OclConfig_gen.h
@@ -132,22 +132,6 @@ size_t inline generate<Algorithm::RANDOM_X>(Threads<OclThreads> &threads, const 
 #endif
 
 
-#ifdef XMRIG_ALGO_ASTROBWT
-template<>
-size_t inline generate<Algorithm::ASTROBWT>(Threads<OclThreads>& threads, const std::vector<OclDevice>& devices)
-{
-    size_t count = 0;
-
-    if (!threads.isExist(Algorithm::ASTROBWT_DERO_2)) {
-        count += threads.move(Algorithm::kASTROBWT_DERO_2, OclThreads(devices, Algorithm::ASTROBWT_DERO_2));
-    }
-
-    count += generate(Algorithm::kASTROBWT, threads, Algorithm::ASTROBWT_DERO, devices);
-    return count;
-}
-#endif
-
-
 #ifdef XMRIG_ALGO_KAWPOW
 template<>
 size_t inline generate<Algorithm::KAWPOW>(Threads<OclThreads>& threads, const std::vector<OclDevice>& devices)