+++ src/backend/opencl/OclThread.h
@@ -75,20 +75,6 @@ public:
     }
 #   endif
 
-#   ifdef XMRIG_ALGO_ASTROBWT
-    OclThread(uint32_t index, uint32_t intensity, uint32_t threads) :
-        m_fields(4),
-        m_threads(threads, -1),
-        m_index(index),
-        m_memChunk(0),
-        m_stridedIndex(0),
-        m_unrollFactor(1),
-        m_worksize(1)
-    {
-        setIntensity(intensity);
-    }
-#   endif
-
 #   ifdef XMRIG_ALGO_KAWPOW
     OclThread(uint32_t index, uint32_t intensity, uint32_t worksize, uint32_t threads) :
         m_fields(8),
@@ -127,7 +113,6 @@ private:
     enum Fields {
         STRIDED_INDEX_FIELD,
         RANDOMX_FIELDS,
-        ASTROBWT_FIELDS,
         KAWPOW_FIELDS,
         FIELD_MAX
     };