+++ src/base/crypto/Algorithm.h
@@ -82,8 +82,6 @@ public:
         AR2_CHUKWA      = 0x61130000,   // "argon2/chukwa"    Argon2id (Chukwa).
         AR2_CHUKWA_V2   = 0x61140000,   // "argon2/chukwav2"  Argon2id (Chukwa v2).
         AR2_WRKZ        = 0x61120000,   // "argon2/wrkz"      Argon2id (WRKZ)
-        ASTROBWT_DERO   = 0x41000000,   // "astrobwt"         AstroBWT (Dero)
-        ASTROBWT_DERO_2 = 0x41110000,   // "astrobwt/v2"      AstroBWT (Dero HE)
         KAWPOW_RVN      = 0x6b0f0000,   // "kawpow/rvn"       KawPow (RVN)
 
         RX_XLA          = 0x721211ff,   // "panthera"         Panthera (Scala2).
@@ -99,7 +97,6 @@ public:
         CN_FEMTO        = 0x63110000,
         RANDOM_X        = 0x72000000,
         ARGON2          = 0x61000000,
-        ASTROBWT        = 0x41000000,
         KAWPOW          = 0x6b000000,
         GHOSTRIDER      = 0x6c000000
     };
@@ -163,12 +160,6 @@ public:
     static const char *kAR2_WRKZ;
 #   endif
 
-#   ifdef XMRIG_ALGO_ASTROBWT
-    static const char *kASTROBWT;
-    static const char *kASTROBWT_DERO;
-    static const char* kASTROBWT_DERO_2;
-#   endif
-
 #   ifdef XMRIG_ALGO_KAWPOW
     static const char *kKAWPOW;
     static const char *kKAWPOW_RVN;
@@ -205,14 +196,7 @@ public:
     inline uint32_t minIntensity() const                    { return ((m_id == GHOSTRIDER_RTM) ? 8 : 1); };
     inline uint32_t maxIntensity() const                    { return isCN() ? 5 : ((m_id == GHOSTRIDER_RTM) ? 8 : 1); };
 
-    inline size_t l3() const
-    {
-#       ifdef XMRIG_ALGO_ASTROBWT
-        return m_id != ASTROBWT_DERO ? l3(m_id) : 0x100000 * 20;
-#       else
-        return l3(m_id);
-#       endif
-    }
+    inline size_t l3() const                                { return l3(m_id); }
 
     inline bool operator!=(Algorithm::Id id) const          { return m_id != id; }
     inline bool operator!=(const Algorithm &other) const    { return !isEqual(other); }