+++ src/base/crypto/Coin.cpp
@@ -49,8 +49,6 @@ static const CoinInfo coinInfo[] = {
     { Algorithm::RX_0,            "XMR",      "Monero",       120,    1000000000000,  YELLOW_BG_BOLD( WHITE_BOLD_S " monero  ") },
     { Algorithm::CN_R,            "SUMO",     "Sumokoin",     240,    1000000000,     BLUE_BG_BOLD(   WHITE_BOLD_S " sumo    ") },
     { Algorithm::RX_ARQ,          "ARQ",      "ArQmA",        120,    1000000000,     BLUE_BG_BOLD(   WHITE_BOLD_S " arqma   ") },
-    { Algorithm::ASTROBWT_DERO,   "DERO",     "DERO",         0,      0,              BLUE_BG_BOLD(   WHITE_BOLD_S " dero    ") },
-    { Algorithm::ASTROBWT_DERO_2, "DERO_HE",  "DERO_HE",      0,      0,              BLUE_BG_BOLD(   WHITE_BOLD_S " dero_he ") },
     { Algorithm::RX_GRAFT,        "GRFT",     "Graft",        120,    10000000000,    BLUE_BG_BOLD(   WHITE_BOLD_S " graft   ") },
     { Algorithm::RX_KEVA,         "KVA",      "Kevacoin",     0,      0,              MAGENTA_BG_BOLD(WHITE_BOLD_S " keva    ") },
     { Algorithm::KAWPOW_RVN,      "RVN",      "Ravencoin",    0,      0,              BLUE_BG_BOLD(   WHITE_BOLD_S " raven   ") },