+++ src/base/net/stratum/Job.cpp
@@ -169,10 +169,6 @@ int32_t xmrig::Job::nonceOffset() const
    auto f = algorithm().family();
    if (f == Algorithm::KAWPOW)     return 32;
    if (f == Algorithm::GHOSTRIDER) return 76;
-
-   auto id = algorithm().id();
-   if (id == Algorithm::ASTROBWT_DERO_2) return 44;
-
    return 39;
 }